change ui and add fonts

+++ Engine/src/Platform/OpenGL/OpenGLVertexArray.cpp
@@ -74,7 +74,7 @@ namespace Platform::OpenGL
                                   GetGLenumFromShaderDataType(element.Type),
                                   element.Normalized ? GL_TRUE : GL_FALSE,
                                   layout.GetStride(),
-                                  reinterpret_cast<const void *>(element.Offset));
+                                  reinterpret_cast<const void *>((uint64_t)element.Offset));
             index++;
         }
         m_VertexBuffers.push_back(vertexBuffer);
